add readme on location of /sso routes
